requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
numpy==2.2.4
scikit-learn==1.6.1
transformers>=4.36.0
torch==2.6.0
soundfile==0.12.1
spotipy==2.23.0
youtube-search-python==1.6.6
tokenizers>=0.21.1
